+++ src/pages/Create.tsx
@@ -1,0 +1,41 @@
+import { FormEvent, useState } from 'react'
+import usePosts from '../hooks/usePost'
+import classes from './Home.module.css'
+
+const Create = () => {
+  const { posts, isLoading, isSubmitting, createPost } = usePosts()
+  const [newTitle, setNewTitle] = useState<string>('')
+  const [newBody, setNewBody] = useState<string>('')
+
+  const handleSubmit = async (e: FormEvent) => {
+    e.preventDefault()
+
+    try {
+      await createPost(newTitle, newBody)
+
+      setNewTitle('')
+      setNewBody('')
+    } catch (err) {
+      console.error(err)
+    }
+  }
+
+  if (isLoading) return <h1>Loading...</h1>
+
+  return (
+    <div className={classes.container}>
+      <form className={classes.postForm} onSubmit={handleSubmit}>
+        <label>Title</label>
+        <input type="text" value={newTitle} onChange={(e) => setNewTitle(e.target.value)} required />
+        <label>Body</label>
+        <input type="text" value={newBody} onChange={(e) => setNewBody(e.target.value)} required />
+
+        <button type="submit" disabled={isSubmitting}>
+          {isSubmitting ? 'Submitting...' : 'Submit'}
+        </button>
+      </form>
+    </div>
+  )
+}
+
+export default Create
